docs(swagger): core more localai/openai endpoints
